requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
airo-dataset-tools @ file:///home/matt/System/airo-mono/airo-dataset-tools
airo-drake==0.0.6
airo-models==0.0.10
airo-tulip==0.1.0
-e git+https://github.com/airo-ugent/airo-mono.git@b9d8182afbc479f93df35266a1ead2e8bba02f81#egg=airo_camera_toolkit&subdirectory=airo-camera-toolkit
-e git+https://github.com/airo-ugent/airo-mono.git@b9d8182afbc479f93df35266a1ead2e8bba02f81#egg=airo_robots&subdirectory=airo-robots
-e git+https://github.com/airo-ugent/airo-mono.git@b9d8182afbc479f93df35266a1ead2e8bba02f81#egg=airo_spatial_algebra&subdirectory=airo-spatial-algebra
-e git+https://github.com/airo-ugent/airo-mono.git@b9d8182afbc479f93df35266a1ead2e8bba02f81#egg=airo_typing&subdirectory=airo-typing
albumentations==1.4.0
annotated-types==0.7.0
ansitable==0.11.4
anyio==4.8.0
argcomplete==3.5.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
boto3==1.35.98
botocore==1.35.98
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colored==2.2.4
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7
cycler==0.12.1
cyclonedds @ git+https://github.com/eclipse-cyclonedds/cyclonedds-python@f8b2b1d56e28f07339844c8d7c763d8842df4ff2
dacite==1.7.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.9
distlib==0.3.9
dnspython==2.7.0
drake==1.26.0
embreex==2.17.7.post5
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fiftyone==1.2.0
fiftyone-brain==0.18.2
fiftyone_db==1.1.7
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
future==1.0.0
glob2==0.7
graphql-core==3.2.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
Hypercorn==0.17.3
hyperframe==6.0.1
identify==2.6.5
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflate64==1.0.1
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.43.0
loguru==0.7.3
lxml==5.3.0
manifold3d==3.0.1
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.1
mongoengine==0.29.1
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
multivolumefile==0.2.3
munch @ file:///home/conda/feedstock_root/build_artifacts/munch_1734441240299/work
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.0
numpy-quaternion==2024.0.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
-e git+https://github.com/NVlabs/curobo.git@2fbffc35225398cf9d5f382804faa9de2608753b#egg=nvidia_curobo
open3d==0.19.0
opencv-contrib-python==4.8.1.78
opencv-python==4.10.0.84
opencv-python-headless==4.8.1.78
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
pprintpp==0.4.0
pre_commit==4.0.1
priority==2.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==0.22.0
pyarrow==18.1.0
pybcj==1.0.3
pybind11==2.13.6
pycocotools==2.0.8
pycollada==0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydash==8.0.4
pydot==3.0.4
Pygments==2.19.1
pykalman==0.9.7
pymongo==4.9.2
pyparsing==3.2.1
pyppmd==1.1.1
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
pyserial==3.5
pysoem==1.1.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.0.3
pyzstd==0.16.2
qudida==0.0.4
rarfile==4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rerun-sdk==0.21.0
retrying==1.3.4
rich==13.9.4
rich-click==1.8.5
rpds-py==0.22.3
Rtree==1.3.0
s3transfer==0.10.4
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentencepiece==0.2.0
setuptools-scm==8.1.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
spatialmath-python==1.1.13
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.45.2
strawberry-graphql==0.258.0
svg.path==6.3
sympy==1.13.1
tabulate==0.9.0
taskgroup==0.2.2
tenacity==9.0.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tomli==2.2.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.3
triton==3.1.0
types-Pillow==10.2.0.20240822
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.63
ultralytics-thop==2.0.14
universal-analytics-python3==1.1.1
ur-analytic-ik==0.0.7
ur_rtde==1.6.0
urllib3==2.3.0
vhacdx==0.0.8.post1
virtualenv==20.28.1
voxel51-eta==0.13.1
warp-lang==1.5.1
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xatlas==0.0.9
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
yourdfpy==0.0.56
zipp==3.21.0
